Renamed package .internal.session.room.annotation to xx.relation
