+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/session/room/Room.kt
@@ -19,7 +19,7 @@ package im.vector.matrix.android.api.session.room
 import androidx.lifecycle.LiveData
 import im.vector.matrix.android.api.session.room.members.MembershipService
 import im.vector.matrix.android.api.session.room.model.RoomSummary
-import im.vector.matrix.android.api.session.room.model.annotation.RelationService
+import im.vector.matrix.android.api.session.room.model.relation.RelationService
 import im.vector.matrix.android.api.session.room.read.ReadService
 import im.vector.matrix.android.api.session.room.send.SendService
 import im.vector.matrix.android.api.session.room.state.StateService
