+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/session/room/model/relation/RelationContent.kt
@@ -1,4 +1,4 @@
-package im.vector.matrix.android.api.session.room.model.annotation
+package im.vector.matrix.android.api.session.room.model.relation
 
 interface RelationContent {
     val type: String?
